callback system

+++ Sources/helpers/string.c
@@ -1,9 +1,9 @@
 #include"../../Headers/helpers/string.h"
 
-unsigned short      str_to_int(const char * s)
+int     str_to_int(const char * s)
 {
-    unsigned short	i;
-    unsigned short	v;
+    int i;
+    int v;
 
     if (s[0] == '-') return 0;
     for(v = 0; s[i] >= 48  && s[i] <= 57; ++i) {
